yielder: support for-in statements over an Iterator.
This also implements for-in support in generators.  All for-in statements
are rewritten as while loops over an Iterator object; the traditional
  for (p in o) { ... }
is equivalent to the new-style:
  for (p in Iterator(o, true)) { ... }
Once the 'in' object in all for-in statements is an Iterator, the for
can be rewritten as:
  for ($it=Iterator(...);;) {
    try {
      p = $it.next();
    } catch (e) {
      if (e===StopIteration) break;
      throw e;
    }
    ...
  }

Now that all for-in loops have been transformed to 'normal' for loops,
they can also be included in generators -- almost: I need to support
'break' and 'continue' in generators first.