yielder: fix ARRAY_COMP by updating narcissus to remove leading [ from parent.
This commit has been rebased anachronistically.  Originally the ARRAY_COMP
handler worked around the fact that narcissus included the leading [ of
the ARRAY_COMP in the src of the parent.  Then at some point I fixed that
in Narcissus (which is an out-of-order commit in this tree, since Narcissus
has been rebased further up in the history), and this commit removed the
crazy ARRAY_COMP workaround since Narcissus was now sane.